Added typescript code
