+++ web-app/src/components/StylesDictionary.tsx
@@ -1,0 +1,5 @@
+import {CSSProperties} from "react";
+
+export interface StylesDictionary {
+    [Key: string]: CSSProperties;
+}
