+++ web-app/src/components/analytics/ImageFrameCard.tsx
