+++ web-app/src/components/dashboard/InfoBox.tsx
@@ -1,12 +1,26 @@
 import React from "react";
 import PropTypes from "prop-types";
+import CountUp from "react-countup";
 import Paper from "@material-ui/core/Paper";
-import {grey} from "@material-ui/core/colors";
+import {grey, lightBlue} from "@material-ui/core/colors";
+import {StylesDictionary} from "../StylesDictionary";
 
-class DataBox extends React.Component {
+
+const propTypes = {
+    Icon: PropTypes.any, // eslint-disable-line
+    color: PropTypes.string, // eslint-disable-line
+    title: PropTypes.string,
+    value: PropTypes.any,
+    start: PropTypes.number,
+};
+
+type ComponentProps = PropTypes.InferProps<typeof propTypes>;
+
+class InfoBox extends React.Component<ComponentProps, any> {
     render() {
-        const {color, title, value, Icon} = this.props;
-        const styles = {
+        const {color, title, value, Icon, start = 0} = this.props;
+        // @ts-ignore
+        const styles: StylesDictionary = {
             content: {
                 padding: "5px 10px",
                 marginLeft: 90,
@@ -28,7 +42,7 @@ class DataBox extends React.Component {
                 height: 90,
                 width: 90,
                 textAlign: "center",
-                backgroundColor: color
+                backgroundColor: color || lightBlue[700]
             },
             icon: {
                 height: 48,
@@ -42,14 +56,13 @@ class DataBox extends React.Component {
             <Paper>
         <span style={styles.iconSpan}>
           <div style={{color: "white"}}>
-            <Icon style={styles.icon} color="white"/>
+            <Icon style={styles.icon}/>
           </div>
         </span>
                 <div style={styles.content}>
                     <span style={styles.text}>{title}</span>
                     <div style={styles.number}>
-                        <span style={styles.text}>{value}</span>
-
+                        <CountUp start={start || 0} end={value} separator="," duration={1}/>
                     </div>
                 </div>
             </Paper>
@@ -57,11 +70,5 @@ class DataBox extends React.Component {
     }
 }
 
-DataBox.propTypes = {
-    Icon: PropTypes.any, // eslint-disable-line
-    color: PropTypes.string,
-    title: PropTypes.string,
-    value: PropTypes.string
-};
 
-export default DataBox;
+export default InfoBox;
