+++ web-app/src/index.tsx
@@ -6,10 +6,9 @@ import "font-awesome/css/font-awesome.css";
 import TimeAgo from 'javascript-time-ago'
 
 import en from 'javascript-time-ago/locale/en'
-import ru from 'javascript-time-ago/locale/ru'
 
 TimeAgo.addDefaultLocale(en)
-TimeAgo.addLocale(ru)
+
 
 require("./favicon.ico");
