--- web-app/src/pages/NotFoundPage.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const NotFoundPage = () => {
-    return (
-        <div>
-            <div>404 Page Not Found</div>
-        </div>
-    );
-};
-
-export default NotFoundPage;
