+++ web-app/src/routes.tsx
@@ -1,12 +1,11 @@
 import React from "react";
 import {BrowserRouter as Router, Route, Switch} from "react-router-dom";
 import App from "./pages/App";
-import Login from "./pages/LoginPage";
+
 
 export default (
     <Router>
         <Switch>
-            <Route exact path="/login" component={Login}/>
             <Route path="/" component={App}/>
         </Switch>
     </Router>
