--- web/index.html
@@ -1,14 +1,0 @@
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="">
-
-  </head>
-  <body>
-  <h2><a><a href="/app/"/>Go to ACE controller app</a></h2>
-  <h2><a><a href="/api/v1/ui"/>Go to ACE api documentation</a></h2>
-  </body>
-</html>
